Create README.md
# Quantum Computing and Go Game Optimization Project

## Description

This repository contains quantum computing experiments, specifically focused on the implementation and optimization of Ising models for the game of Go using D-Wave technology and the dimod library.

## Experiments

The main experiment is located in the AtomicGo QUBO.ipynb file. This Jupyter notebook details the process and results of the experiment, including the formulation of the problem as an Unconstrained Binary Quadratic Optimization (QUBO) and its resolution using a quantum annealer based on the Ising model.

## Technologies Used

- Python: Main programming language for the experiments.
- D-Wave: Quantum computing platform used for Ising model optimization.
- dimod: Library for defining binary quadratic models.
- Jupyter Notebook: For documentation and visualization of the experiments.

## How to Contribute

If you are interested in contributing to this project, please:

1. Fork the repository.
2. Create a new branch in your fork.
3. Make your changes or additions.
4. Submit a pull request to merge your changes into the main repository.

## License

This project is under the MIT license. See the LICENSE file for more details.